fix: 收紧搜索来源去重边界
